Fake-News-Detection-Web-App-main/requirements.txt
absl-py==2.1.0
altair==5.2.0
astunparse==1.6.3
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.3.1
docopt==0.6.2
feedfinder2==0.0.4
feedparser==6.0.11
flatbuffers==23.5.26
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.42
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-generativeai==0.3.2
google-pasta==0.2.0
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-status==1.62.0
idna==3.6
imageio==2.34.0
importlib-metadata==7.0.1
ipython==8.12.3
jieba3k==0.35.1
Jinja2==3.1.3
joblib==1.2.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==2.15.0
libclang==16.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.2.0
nbconvert==7.16.1
newspaper3k==0.2.8
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.1
pillow==10.2.0
pipreqs==0.5.0
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==15.0.0
PyAudio==0.2.14
pydeck==0.8.1b0
Pygments==2.17.2
pypiwin32==223
python-dateutil==2.8.2
python-dotenv==1.0.1
pyttsx3==2.90
pytz==2024.1
pywin32==306
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.18.0
rsa==4.9
scipy==1.12.0
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
SpeechRecognition==3.10.1
streamlit==1.31.1
tenacity==8.2.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tinysegmenter==0.3
toml==0.10.2
toolz==0.12.1
torch==2.2.1
tornado==6.4
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
validators==0.22.0
watchdog==4.0.0
wrapt==1.16.0
yarg==0.1.9
zipp==3.17.0
